# target/generated-sources/swagger.json
{"swagger":"2.0","info":{"description":"API for fetching hoge related information","version":"1.0.0","title":"Hoge User API"},"host":"localhost","basePath":"/","tags":[{"name":"hoge-user-controller","description":"Hoge User Controller"}],"paths":{"/api/hoge-user":{"get":{"tags":["hoge-user-controller"],"summary":"findAll","operationId":"findAllUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"Page number of the results you want to fetch","required":false,"type":"integer","default":1,"format":"int32","allowEmptyValue":false},{"name":"size","in":"query","description":"Total amount of elements on a specific page","required":false,"type":"integer","default":10,"format":"int32","allowEmptyValue":false}],"responses":{"200":{"description":"Success","headers":{"X-Users-Total":{"type":"integer","format":"int64","description":"Total number of users that can be found"}},"schema":{"type":"array","items":{"$ref":"#/definitions/HogeUser"}}},"400":{"description":"Bad request","schema":{"type":"array","items":{"$ref":"#/definitions/ApiError"}}},"500":{"description":"Internal Server Error","schema":{"type":"array","items":{"$ref":"#/definitions/ApiError"}}}},"deprecated":false},"post":{"tags":["hoge-user-controller"],"summary":"save","operationId":"saveUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/HogeUserInput"}}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/HogeUser"}}},"400":{"description":"Bad request","schema":{"type":"array","items":{"$ref":"#/definitions/ApiError"}}},"500":{"description":"Internal Server Error","schema":{"type":"array","items":{"$ref":"#/definitions/ApiError"}}}},"deprecated":false}},"/api/hoge-user/{id}":{"get":{"tags":["hoge-user-controller"],"summary":"findOne","operationId":"findOneUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"The unique identifier of the hoge you prefer to fetch","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/HogeUser"}}},"400":{"description":"Bad request","schema":{"type":"array","items":{"$ref":"#/definitions/ApiError"}}},"500":{"description":"Internal Server Error","schema":{"type":"array","items":{"$ref":"#/definitions/ApiError"}}}},"deprecated":false}}},"definitions":{"ApiError":{"type":"object","properties":{"codes":{"type":"array","items":{"type":"string"}},"message":{"type":"string"}},"title":"ApiError"},"HogeUser":{"type":"object","properties":{"dateOfBirth":{"type":"string","format":"date"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int64"},"lastName":{"type":"string"},"middleName":{"type":"string"},"siblings":{"type":"integer","format":"int32"}},"title":"HogeUser"},"HogeUserInput":{"type":"object","properties":{"dateOfBirth":{"type":"string","format":"date"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"siblings":{"type":"integer","format":"int32"}},"title":"HogeUserInput"}}}